src/exam_prep_buddy/tools/multimodal_tool.py
from crewai.tools import BaseTool
from typing import Type
from pydantic import BaseModel, Field
from PIL import Image

class MultimodalInput(BaseModel):
    image_path: str = Field(..., description="Path to the image/diagram file.")
    prompt: str = Field(..., description="Prompt for the multimodal LLM.")

class MultimodalTool(BaseTool):
    name: str = "Multimodal LLM Tool"
    description: str = "Processes images/diagrams and generates visual content using a multimodal LLM."
    args_schema: Type[BaseModel] = MultimodalInput

    def _run(self, image_path: str, prompt: str) -> str:
        # Placeholder: Integrate with your multimodal LLM here
        # For now, just return a stub
        return f"Processed {image_path} with prompt: {prompt} (stub output)"
