src/exam_prep_buddy/tools/pyq_scraper.py
from crewai.tools import BaseTool
from typing import Type
from pydantic import BaseModel, Field
import requests
from bs4 import BeautifulSoup

class PYQScraperInput(BaseModel):
    course_code: str = Field(..., description="Course code to search for PYQs.")

class PYQScraperTool(BaseTool):
    name: str = "PYQ Scraper Tool"
    description: str = "Scrapes previous year question papers from vhelpcc.com for a given course code."
    args_schema: Type[BaseModel] = PYQScraperInput

    def _run(self, course_code: str) -> str:
        # Placeholder: Implement scraping logic for https://www.vhelpcc.com/pyqs
        # For now, just return a stub
        return f"Scraped PYQs for course code: {course_code} (stub output)"
